fix(camera,cockpit): Add missing activity-compose dep, fix Map zoom type
- CameraAvanue: add androidx.activity.compose dependency for
  rememberLauncherForActivityResult and ActivityResultContracts
- Cockpit ContentRenderer: convert Float zoomLevel to Int for
  OpenStreetMap embed URL (zoom levels are integer 1-18)

+++ Modules/Cockpit/src/androidMain/kotlin/com/augmentalis/cockpit/content/ContentRenderer.kt
@@ -149,7 +149,7 @@ fun ContentRenderer(
             is FrameContent.Map -> MapContentRenderer(
                 latitude = content.latitude,
                 longitude = content.longitude,
-                zoomLevel = content.zoomLevel,
+                zoomLevel = content.zoomLevel.toInt(),
                 modifier = Modifier.fillMaxSize()
             )
